Add files via upload
This code needs some cleaning up. E.g. some of the code for plotting is most likely redundant, but it was made in a haste.  In addition some of the variables are used in several object, but instead of fetching them from the other object they are defined several times in their respective objects, making some of the code redundant.